study/requirements.txt
openai>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
flask>=2.0.0
flask-cors>=4.0.0
requests>=2.25.0
pyperclip>=1.8.2